refactor: some refactors
