+++ src/components/restClient/types.ts
@@ -1,15 +1,26 @@
 import type { HttpMethod } from '@/lib/restClient/restClient';
 
+export type ApiResult = Record<string, unknown> | { error: string };
+
 export type FormValues = {
   method: HttpMethod;
-  URL: string;
+  url: string;
   body: string;
+  headers: Header[];
 };
-
-export type ApiResult = Record<string, unknown> | { error: string };
 
 export type ResponseInfo = {
   status: number | null;
   statusText: string;
   duration: number | null;
+};
+
+export type Header = {
+  key: string;
+  value: string;
+  enabled?: boolean;
+};
+
+export type HeadersFormValues = {
+  headers: Header[];
 };
