--- src/lib/store/slice/counterSlice.ts
@@ -1,19 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit';
-
-const initialState = { value: 5 };
-
-const counterSlice = createSlice({
-  name: 'counter',
-  initialState,
-  reducers: {
-    increment: (state) => {
-      state.value += 1;
-    },
-    decrement: (state) => {
-      state.value -= 1;
-    },
-  },
-});
-
-export const { increment, decrement } = counterSlice.actions;
-export default counterSlice.reducer;
